some changes and file organization
