+++ package.json
@@ -15,7 +15,6 @@
     "react": "^17.0.2",
     "react-bootstrap": "^2.8.0",
     "react-dom": "^17.0.2",
-    "react-pdf": "^7.3.3",
     "react-router-dom": "^6.2.2",
     "react-scripts": "^5.0.0",
     "react-table": "^7.7.0",
