--- src/services/service.js
@@ -1,12 +1,0 @@
-/*const serviceResponse = () => (
-    {
-        data: sheetData,
-        loading,
-        error,
-      } = useGoogleSheets({
-        apiKey: process.env.REACT_APP_GOOGLE_API_KEY,
-        sheetId: process.env.REACT_APP_GOOGLE_SHEETS_ID,
-      })
-    );
-
-export default serviceResponse*/
